fix warnings

+++ src/engine/screen.cpp
@@ -1020,14 +1020,14 @@ namespace
                     if ( keepAspectRatio ) {
                         if ( ( static_cast<float>( VITA_FULLSCREEN_WIDTH ) / VITA_FULLSCREEN_HEIGHT ) >= ( static_cast<float>( width_ ) / height_ ) ) {
                             float scale = static_cast<float>( VITA_FULLSCREEN_HEIGHT ) / height_;
-                            destRect.w = width_ * scale;
+                            destRect.w = static_cast<int32_t>( static_cast<float>( width_ ) * scale );
                             destRect.h = VITA_FULLSCREEN_HEIGHT;
                             destRect.x = ( VITA_FULLSCREEN_WIDTH - destRect.w ) / 2;
                         }
                         else {
                             float scale = static_cast<float>( VITA_FULLSCREEN_WIDTH ) / width_;
                             destRect.w = VITA_FULLSCREEN_WIDTH;
-                            destRect.h = height_ * scale;
+                            destRect.h = static_cast<int32_t>( static_cast<float>( height_ ) * scale );
                             destRect.y = ( VITA_FULLSCREEN_HEIGHT - destRect.h ) / 2;
                         }
                     }
